rag-pdf-qa/requirements.txt
google-generativeai>=0.4.0
chromadb>=0.4.24
sentence-transformers>=2.5.1
pdfplumber>=0.11.0
python-docx>=1.1.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.1
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
